requirements.txt
# 核心深度学习框架
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0

# BERT相关
sentence-transformers>=2.2.0

# 向量检索
faiss-cpu>=1.7.4

# 科学计算
numpy>=1.24.0
scipy>=1.10.0

# 数据处理
datasets>=2.12.0
pandas>=2.0.0

# 可视化和日志
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0  # 可选

# 配置管理
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# 工具
tqdm>=4.65.0
rich>=13.0.0